# Conflicts:
#	src/main/java/org/example/Course.java

+++ src/main/java/org/example/FileHandler.java
@@ -16,16 +16,27 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 public class FileHandler {
-    private static String createPath(String nameAndExt){
+    public static String getDefaultPath(String nameAndExt){
         Path currentPath = FileSystems.getDefault().getPath("");
         String currentName = currentPath.toAbsolutePath().toString();
         String filePath = currentName + "\\src\\CourseLists\\";
         return filePath + nameAndExt;
     }
 
-    public static void saveList(CourseList courses, String fileName){
-        String filePath = createPath(fileName + ".json");
+    /**
+     * @param courses is the list of courses to be saved
+     * @param filePath the absolute path to the file where the list is to be saved
+     * @param overWrite if true, an existing file with that specified path can be deleted and replaced
+     *                  if false, finding an existing file with that name results in returning false
+     * @return true if the file was successfully saved
+     */
+    public static boolean saveList(CourseList courses, String filePath, boolean overWrite){
         File outFile = new File(filePath);
+
+        if(outFile.exists() && !overWrite){
+            return false;
+        }
+
         FileOutputStream outStream = null;
         try{
             outStream = new FileOutputStream(outFile);
@@ -37,23 +48,25 @@ public class FileHandler {
         for(Course course : courses.getCourses()){
             writer.write(course);
         }
+        return true;
     }
 
-    public static CourseList loadList(String fileName){
-        String filePath = createPath(fileName + ".json");
+    /**
+     * @param filePath the absolute path to the file to be loaded
+     * @return the list of courses found in the specified json file
+     * @throws FileNotFoundException
+     */
+    public static CourseList loadList(String filePath) throws FileNotFoundException {
         FileInputStream inStream = null;
-        try {
-            inStream = new FileInputStream(filePath);
-        }catch(FileNotFoundException e){
-            e.printStackTrace();
-        }
+        inStream = new FileInputStream(filePath);
+
         Scanner scan = new Scanner(inStream);
         scan.useDelimiter(",");
         String workingStr = "";
         /* workingStr will represent some fraction of the input file, starting from a '{'
-        *  When an object has been read from the json file, the beginning of workingStr advances
-        *   to the start of the next object
-        * */
+         *  When an object has been read from the json file, the beginning of workingStr advances
+         *   to the start of the next object
+         * */
         int openBrackets = 0;
         int currentIndex = 0;
 
@@ -64,7 +77,7 @@ public class FileHandler {
 
             workingStr += scan.next();
             if(scan.hasNext()){
-                workingStr += ','; //add a comma every time except the last time
+                workingStr += ','; //add a comma every time except the last time to preserve indexing
             }
 
             for(; currentIndex < workingStr.length(); currentIndex++){
@@ -76,21 +89,23 @@ public class FileHandler {
                 }
 
 
-                if(openBrackets == 0){//we've reached the end of an object
+                if(openBrackets == 0){//currentIndex is the end of an object
                     String objectString = workingStr.substring(0,currentIndex+1);
                     loadedCourse = (Course) JsonReader.jsonToJava(objectString);
+                    loadedList.addCourse(loadedCourse);
+
                     workingStr = workingStr.substring(currentIndex+1);
                     currentIndex = -1; //the for statement will set it back to 0
                 }
             }
-            loadedList.addCourse(loadedCourse);
+
         }
 
         return loadedList;
     }
 
     public static CourseList loadCatalog(){
-        String catalogFilePath = createPath("catalog.xlsx");
+        String catalogFilePath = getDefaultPath("catalog.xlsx");
         CourseList catalog = new CourseList();
 
         try {
@@ -187,15 +202,169 @@ public class FileHandler {
                 }
 
                 ArrayList<String> prereqs = null;
+                Cell preReqCell = thisRow.getCell(19);
+                if(preReqCell != null){
+                    prereqs = parsePrereqs(preReqCell.getStringCellValue());
+                }
 
                 Course thisCourse = new Course(name, code, meetTimes, isFall, description, location, professor, credits, prereqs);
                 catalog.addCourse(thisCourse);
-
-                //TODO Prerequisites!!!
             }
         }catch(Exception e){
             e.printStackTrace();
         }
         return catalog;
     }
+
+    public static boolean saveFutureList(ArrayList<CourseList> futureLists, String filename, boolean overWrite){
+        //algorithm: make a new directory based on filename, and save each courselist inside that directory with the existing
+
+        String filePath = getDefaultPath(filename);
+        File directory = new File(filePath);
+
+        if(directory.exists()) {
+            if(!overWrite){
+                return false;
+            }else {
+                File[] dirContents = directory.listFiles();
+                for(File f : dirContents){
+                    f.delete();
+                }
+                directory.delete();//can only delete an empty directory
+            }
+        }
+        if(!directory.mkdir()){//the directory should exist now
+            return false;
+        }
+
+        for(int x = 0; x < futureLists.size(); x++){
+            String thisPath = filePath + "\\" + filename + "-" + x + ".json";
+            if(!saveList(futureLists.get(x),thisPath,overWrite)){//this is fine. It will work once my last pull request is approved
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public static ArrayList<CourseList> loadFutureList(String filePath) throws FileNotFoundException{
+        ArrayList<CourseList> result = new ArrayList<>();
+
+        File directory = new File(filePath);
+        if(!directory.exists()){
+            throw new FileNotFoundException("We could not find the file: " + filePath);
+        }
+
+        File[] courseLists = directory.listFiles();
+        for(int x = 0; x < courseLists.length; x++){
+            String thisPath = courseLists[x].getPath();
+            CourseList thisList = loadList(thisPath);
+            result.add(thisList);
+        }
+
+        return result;
+    }
+
+    /*
+     *  Dr. Hutchins told me to focus on a simple proof of concept, then move on to something else,
+     *  so this method only works for a single prereq class, and it must be a course code that takes up the first eight chars of the string
+     */
+    private static ArrayList<String> parsePrereqs(String reqsFromFile){
+        try {
+            ArrayList<String> result = new ArrayList<>();
+
+            boolean validCode = true;
+            for(int x = 0; x < 8 && x < reqsFromFile.length() && validCode; x++){
+                char thisChar = reqsFromFile.charAt(x);
+                if(x < 4 && !((thisChar >= 'a' && thisChar <= 'z') || (thisChar >= 'A' && thisChar <= 'Z'))){
+                    validCode = false; //if any of the first four characters is not a letter, it's no good
+                }
+                if(x == 4 && reqsFromFile.charAt(x) != ' '){
+                    validCode = false;
+                }
+                if(x > 4 && !(thisChar >= '0' && thisChar <= '9')){
+                    validCode = false;
+                }
+            }
+            if(validCode) {
+                result.add(reqsFromFile.substring(0, 4) + reqsFromFile.substring(5, 8));
+                return result;
+            }
+            else{return null;}
+            /*
+            ArrayList<String> pieces = new ArrayList<>();
+            Scanner scan = new Scanner(reqsFromFile);
+            while(scan.hasNext()){
+                pieces.add(scan.next());
+            }
+
+            Map<String, String> abbrs = new HashMap<String, String>();
+            abbrs.put("AS","ASTR");
+            abbrs.put("PH","PHYS");
+            abbrs.put("CHM","CHEM");
+            abbrs.put("CM","COMM");
+            abbrs.put("CP","COMP");
+            abbrs.put("MT","Math");
+            abbrs.put("EN","ENGR");
+            abbrs.put("ENG","ENGR");
+            abbrs.put("CH","CHEM");
+            abbrs.put("ME","MECE");
+            abbrs.put("FN","FNCE");
+            abbrs.put("IN","INBS");
+            abbrs.put("M","MATH");
+            abbrs.put("MA","MATH");
+
+            //replace abbreviations with actual department codes
+            for(int x = 0; x < pieces.size(); x++){
+                String thisPiece = pieces.get(x);
+                if(abbrs.containsKey(thisPiece)){
+                    pieces.set(x,abbrs.get(thisPiece));//replace the abbreviation with the full code
+                    continue;
+                }
+            }
+
+            //add course codes to result as appropriate
+            for(int x = 0; x < pieces.size(); x++) {
+                String thisPiece = pieces.get(x);
+                if(thisPiece.length() >= 3){
+                    boolean isNumCode = true;
+                    for(int i = 0; i < 3; i++){
+                        if(!((int)thisPiece.charAt(i) >= 48 && (int)thisPiece.charAt(i) <= 57)){//one of the first 3 chars is non-numeric
+                            isNumCode = false;
+                        }
+                    }
+                    if(isNumCode && x > 0) {
+                        result.add(pieces.get(x - 1) + thisPiece.substring(0, 3));
+                    }
+                }
+
+
+                for(int y = 0; y < pieces.get(x).length(); y++){
+
+                }
+            }
+            */
+        }catch(Exception e){
+            e.printStackTrace();
+            return null;
+        }
+    }
+    /*
+     *unusual department codes:
+     * ART is only 3 letters
+     * AS is short for ASTR
+     * PH is short for PHYS
+     * CHM is short for CHEM
+     * CM is short for COMM
+     * CP is short for COMP
+     * MT is short for MATH
+     * ENG is short for ENGR
+     * EN is short for ENGR
+     * CH is short for CHEM
+     * ME is short for MECE
+     * FN is short for FNCE
+     * IN is short for INBS
+     * MA is short for MATH
+     * M is short for MATH
+     */
+
 }
